TimeComplexity: O(n)
Given two integer array A and B of size N and M respectively. Your task is to find all the common elements in both the array.
NOTE:
Each element in the result should appear as many times as it shows in both arrays.
The result can be in any order.